Restructureing the whole project
added correct HPP and methods in CPP, implemented some of the combat system.
