+++ EscapeProtocol.vcxproj.filters
@@ -18,5 +18,19 @@
     <ClInclude Include="maze.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="JRPG.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="JRPG.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="main.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="maze.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
