+++ JRPG.hpp
@@ -1,0 +1,6 @@
+#ifndef JRPG_HPP
+#define JRPG_HPP
+
+int getValidatedInput(int minRange, int maxRange);
+
+#endif
